try to parse Event Property values as edn
If one logs a nested context map where the keys are vectors or maps
then the logging framework calls to string on them and the event
property ends up being a end string.
In the layout class we try to parse that back to produce richer outputs
that are able to produce nested json for example.
If parsing the value as edn fails we call back to the original input.